requirements.txt
pyinstaller==5.7.0
pillow==9.4.0
pypdf2==3.0.1
cryptography==39.0.1
jinja2==3.1.2
argparse==1.4.0
